fix: add throw error if both files exists _rename.js

+++ src/fs/rename.js
@@ -8,11 +8,27 @@ const target = path.resolve(__dirname, 'files', 'properFilename.md');
 
 const errorMessage = 'FS operation failed';
 
+class isFileExistError extends Error {
+  constructor(msg) {
+    super(msg);
+    this.name = 'isFileExistError';
+  }
+}
+
 const rename = async () => {
   try {
-    await fs.rename(original, target);
-  } catch (error) {
-    throw new Error(errorMessage);
+    await fs.access(target);
+    throw new isFileExistError('');
+  } catch (err) {
+    if (err.name === 'isFileExistError') {
+      throw new Error(errorMessage);
+    } else {
+      try {
+        await fs.rename(original, target);
+      } catch (error) {
+        throw new Error(errorMessage);
+      }
+    }
   }
 };
